requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
affine==2.4.0
    # via rasterio
anyio==3.7.0
    # via starlette
attrs==22.2.0
    # via
    #   fiona
    #   rasterio
certifi==2022.12.7
    # via
    #   fiona
    #   pyproj
    #   rasterio
    #   requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   click-plugins
    #   cligj
    #   fiona
    #   geomet
    #   rasterio
    #   sentinelsat
click-plugins==1.1.1
    # via
    #   fiona
    #   rasterio
cligj==0.7.2
    # via
    #   fiona
    #   rasterio
exceptiongroup==1.1.1
    # via anyio
fastapi==0.95.2
    # via -r requirements.in
fiona==1.9.3
    # via geopandas
geojson==3.0.1
    # via sentinelsat
geomet==1.0.0
    # via sentinelsat
geopandas==0.12.2
    # via -r requirements.in
html2text==2020.1.16
    # via sentinelsat
idna==3.4
    # via
    #   anyio
    #   requests
munch==2.5.0
    # via fiona
numpy==1.24.2
    # via
    #   pandas
    #   rasterio
    #   shapely
    #   snuggs
packaging==23.1
    # via geopandas
pandas==2.0.0
    # via geopandas
pydantic==1.10.8
    # via fastapi
pyparsing==3.0.9
    # via snuggs
pyproj==3.5.0
    # via geopandas
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via pandas
rasterio==1.3.6
    # via -r requirements.in
requests==2.28.2
    # via sentinelsat
sentinelsat==1.2.1
    # via -r requirements.in
shapely==2.0.1
    # via geopandas
six==1.16.0
    # via
    #   geomet
    #   munch
    #   python-dateutil
sniffio==1.3.0
    # via anyio
snuggs==1.4.7
    # via rasterio
starlette==0.27.0
    # via fastapi
tqdm==4.65.0
    # via sentinelsat
typing-extensions==4.6.2
    # via pydantic
tzdata==2023.3
    # via pandas
urllib3==1.26.15
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
